etl/test-requirements.txt
kafka-python==2.0.2
python-snappy==0.6.0
SQLAlchemy==2.0.7
PyMySQL==1.0.2

pytest
docker
pytest-timeout